Moved sidebar data into  so it wasn't taking up a ton of space in the layout file. Fixed some weird bugs with sidebar buttons and the sign-out button not filling to the edge correctly. Applied the 'disabled' property to the correct element on the sidebar so it no longer highlights when hovering over it. Work is beginning on the event page.
