+++ src/routes/home/+layout.ts
@@ -1,0 +1,7 @@
+import { mainSidebarItems } from '$lib/sidebar';
+import type { LayoutLoad } from './$types';
+
+
+export const load: LayoutLoad = async ({ data }) => {
+  return { user: data.user, sidebarItems: mainSidebarItems };
+}
